pyproject.toml
[project]
name = "physae"
version = "0.1.0"
description = "Modular PhysAE project"
authors = [{name = "PhysAE Team"}]
requires-python = ">=3.10"
dependencies = [
    "torch>=2.1.0",
    "fastapi>=0.110",
    "uvicorn[standard]>=0.23",
    "pydantic>=1.10",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "httpx>=0.24",
    "ruff>=0.1.8",
    "black>=23.9",
    "mypy>=1.5",
    "pre-commit>=3.4",
]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "UP", "B", "N"]
ignore = ["E203", "E501"]

[tool.mypy]
python_version = 3.11
warn_unused_ignores = true
warn_return_any = true
warn_unused_configs = true
strict_optional = true
mypy_path = ["src"]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
pythonpath = ["src"]

[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"
